[Refactor] 서버 코드 메서드 별 분리 및 유틸 분리 #214
