+++ src/app/api/auth/check/route.ts
