--- src/app/api/auth/session/index.ts
@@ -1,2 +1,0 @@
-export { POST } from "./post";
-export { DELETE } from "./delete";
