+++ src/app/api/auth/users/delete.ts
@@ -1,8 +1,7 @@
 import { NextRequest } from "next/server";
 import bcrypt from "bcrypt";
 
-import clientPromise from "@/lib/mongodb";
-import { verifyToken } from "@/utils/server/tokenHelper";
+import { verifyToken } from "@/server/utils/tokenHelper";
 import {
   handleErrorResponse,
   handleSuccessResponse,
@@ -12,14 +11,11 @@ import {
   NotFoundError,
   ValidationError,
 } from "@/server/customErrors";
+import { getCollection } from "@/server/db/mongodb";
 
 /** 회원탈퇴 */
 export async function DELETE(request: NextRequest) {
   try {
-    const client = await clientPromise;
-    const db = client.db("wattsup");
-    const collection = db.collection("userdata");
-
     const accessToken = request.cookies.get("accessToken")?.value;
     if (!accessToken) {
       throw new ValidationError(
@@ -39,6 +35,7 @@ export async function DELETE(request: NextRequest) {
       throw new ValidationError("password", "비밀번호를 입력해주세요.");
     }
 
+    const collection = await getCollection("userdata");
     const user = await collection.findOne({ email });
     if (!user) {
       throw new NotFoundError(
