--- src/app/api/auth/users/index.ts
@@ -1,2 +1,0 @@
-export { POST } from "./post";
-export { DELETE } from "./delete";
