+++ src/app/api/auth/users/route.ts
@@ -1,1 +1,2 @@
-export { POST, DELETE } from "./index";
+export { POST } from "./post";
+export { DELETE } from "./delete";
